move baserepo pra providers/ (paridade com baseservice)
baserepo (port generico de persistencia) sai de domain/ e vai pra
providers/, ao lado de baseservice. Justificativa:

- Paridade com baseservice: ambos sao primitivos compartilhados de
  aplicacao, nao tipos de dominio nem use cases.
- O contrato Repository[T] expoe *sqlx.Tx (Commit/Rollback/InitTX) —
  isso e infra, nao dominio. Manter em domain forcava sqlx no toml
  de domain.

Resultado:
- domain.allowed perde sqlx, ganha apenas providers/baserepo.
- providers.allowed ganha sqlx.
- Excecao do query-check no analyser realocada de domain/baserepo
  pra providers/baserepo.

Arch-analyse continua 0 violacoes.
