+++ internal/application/providers/baserepo/baserepo.go
@@ -1,7 +1,9 @@
-// Package baserepo expõe o contrato (port) genérico dos repositórios de
-// domínio. A interface Repository[T] é o conjunto canônico de operações
-// que qualquer adapter de persistência precisa oferecer; o struct concreto
-// que satisfaz essa interface vive em internal/repositories/base_repository.
+// Package baserepo expõe o contrato genérico de repositório de domínio.
+// A interface Repository[T] é o conjunto canônico de operações que qualquer
+// adapter de persistência precisa oferecer; o struct concreto BaseRepo[T]
+// que a satisfaz vive em internal/repositories/base_repository. Convive em
+// providers/ com baseservice (BaseService + ports Logger/IdCreator/Validator)
+// como primitivo cross-cutting de aplicação.
 package baserepo
 
 import (
